modify things and add repositories

+++ src/main/java/com/example/gymweb/RestControllers/UserRestController.java
@@ -1,5 +1,6 @@
 package com.example.gymweb.RestControllers;
 
+import com.example.gymweb.Repositories.UserRepository;
 import com.example.gymweb.Services.UserService;
 import com.example.gymweb.Entities.User;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,6 +16,8 @@ import org.springframework.web.bind.annotation.*;
 public class UserRestController {
     @Autowired
     UserService userService;
+
+
     //code for login into an existing account
     @GetMapping("/login")
     public ResponseEntity<User> login(@RequestParam String email, @RequestParam String password){
@@ -46,12 +49,19 @@ public class UserRestController {
     //delete an user using his ID
     @DeleteMapping("/deleteuser/{id}")
     public ResponseEntity<User> deleteUser(@PathVariable long id) {
-        User user = userService.removeUser(id);
-        if(user != null){
+        if(userService.getUser(id).getId()==-1){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }else{
+            userService.removeUser(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        //userService.removeUser(id);
+
+        /*if(user != null){
             return new ResponseEntity<>(user,HttpStatus.OK);
         }else{
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        }
+        }*/
     }
     //show users based on their ID
     @GetMapping("/users/{id}")
